- bug fix #3004666   - When trying to set texture to nodes from a BMP file, Tulip outputs the following error     Error: number of Planes not 1 in : <BMP file>     This error happens each time I want to set a BMP file as texture.
git-svn-id: https://svn.code.sf.net/p/auber/code/tulip@2528 f5d4c5f9-c943-4916-8a57-bd8605bf961c

+++ library/tulip-ogl/src/GlTextureManager.cpp
@@ -78,7 +78,7 @@ static bool loadBMP(const string &filename, TextureInfo *texture,string &errorMs
   /* skip file size and reserved fields of bitmap file header */
   fseek(file, 8, SEEK_CUR);
   /* get the position of the actual bitmap data */
-  if (!fread(&bfOffBits, sizeof(long int), 1, file))
+  if (!fread(&bfOffBits, sizeof(int), 1, file))
     {
       errorMsg = "Error reading : " + filename;
       //cerr << __PRETTY_FUNCTION__ << ": Error reading " << filename << endl;
